requirements.txt
pillow
google-generativeai
google-api-core
google-auth
requests
